requirements.txt
absl-py==1.4.0
-e git+git@github.com:acl-org/aclpubcheck.git@60d931777a3bb554a6b23d21a4e7b561865e6000#egg=aclpubcheck
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
antlr4-python3-runtime==4.8
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astropy==5.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asyncer==0.0.1
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
basicsr==1.3.3.10
bce-python-sdk==0.8.62
beautifulsoup4==4.10.0
bibtexparser==1.2.0
bleach==3.3.1
blinker==1.4
blis==0.7.4
borb==2.0.24
Brotli==1.0.9
cachetools==4.2.2
catalogue==2.0.6
certifi==2020.6.20
cffi==1.14.5
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.3.1
clean-text==0.6.0
click==8.1.3
colorama==0.4.4
coloredlogs==15.0.1
colorlog==6.4.1
cryptography==36.0.2
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
dataclasses==0.6
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.1.0/de_core_news_sm-3.1.0-py3-none-any.whl
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.2
docker-pycreds==0.4.0
easydict==1.9
editdistance==0.5.3
emoji==1.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.1.0
facexlib==0.1.3.1
fairseq==0.10.0
fastapi==0.80.0
feather==0.1.2
filelock==3.0.12
filetype==1.1.0
flake8==3.9.2
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flatbuffers==2.0.7
fonttools==4.32.0
frozenlist==1.3.1
fsspec==2021.8.1
ftfy==6.1.1
future==0.18.2
fvcore==0.1.3.post20210227
gast==0.3.3
gdown==4.5.1
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.38.1
h11==0.13.0
h2==4.1.0
h5py==3.4.0
hpack==4.0.0
huggingface-hub==0.5.1
humanfriendly==10.0
hydra-core==1.1.0
hyperframe==6.0.1
identify==2.2.14
idna==2.10
ImageHash==4.2.1
imageio==2.9.0
importlib-metadata==3.10.1
iopath==0.1.9
ipykernel==6.0.2
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
jieba==0.42.1
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaitaistruct==0.9
keras==2.11.0
kiwisolver==1.3.1
latexcodec==2.0.1
libclang==15.0.6.1
llvmlite==0.39.0
lmdb==1.2.1
lxml==4.8.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
MechanicalSoup==1.1.0
mediapipe==0.8.7.3
MedPy==0.4.0
mistune==0.8.4
mock==4.0.3
mpmath==1.2.1
multidict==5.1.0
multiprocess==0.70.12.2
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
nibabel==5.0.1
ninja==1.10.2
nltk==3.8.1
nodeenv==1.6.0
notebook==6.4.0
numba==0.56.0
numpy==1.21.6
oauthlib==3.1.1
olefile==0.46
omegaconf==2.1.0
onnx==1.10.1
onnxruntime==1.12.1
openai==0.10.2
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opencv-python-headless==4.6.0.66
openpyxl==3.0.7
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
paddle2onnx==0.8.1
paddlehub==2.1.0
paddlenlp==2.0.8
paddlepaddle==2.1.3
pandas==1.3.1
pandas-stubs==1.2.0.10
pandocfilters==1.4.3
parso==0.8.2
pathtools==0.1.2
pathy==0.6.0
patsy==0.5.2
pdfminer.six==20220319
pdfplumber==0.6.1
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.3.0
plotly==5.10.0
portalocker==2.0.0
pre-commit==2.15.0
preshed==3.0.5
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybtex==0.24.0
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.2
pyDeprecate==0.3.1
pydivert==2.1.0
pyee==7.0.4
pyerfa==2.0.0.1
pyflakes==2.3.1
Pygments==2.9.0
pylatexenc==2.10
PyMatting==1.1.8
PyMuPDF==1.19.6
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyppeteer==0.2.2
pyreadline3==3.4.1
pyrsistent==0.18.0
PySocks==1.7.1
pytesseract==0.3.9
python-barcode==0.13.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
python-slugify==5.0.2
pytorch-lightning==1.4.6
pytz==2021.1
PyWavelets==1.1.1
pywin32==301
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.1.0
qrcode==7.3.1
qtconsole==5.1.1
QtPy==1.9.0
rarfile==4.0
rebiber==1.1.3
regex==2022.10.31
rembg==2.0.24
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacrebleu==1.5.1
sacremoses==0.0.45
scikit-image==0.19.3
scikit-learn==0.24.2
scipy==1.7.3
seaborn==0.11.1
selenium==4.2.0
selenium-wire==4.6.4
Send2Trash==1.7.1
sentry-sdk==1.5.7
seqeval==1.2.2
setproctitle==1.2.2
shellcheck-py==0.7.2.1
shortuuid==1.0.8
SimpleITK==2.2.1
six==1.16.0
smart-open==5.2.1
smmap==4.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
sox==1.4.1
spacy==3.1.2
spacy-legacy==3.0.8
srsly==2.4.1
starlette==0.19.1
statsmodels==0.13.2
subword-nmt==0.3.8
sympy==1.11
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
thinc==8.0.10
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2021.7.30
tika==1.24
tokenizers==0.12.1
toml==0.10.2
torch==1.9.0
torchaudio==0.8.0
torchmetrics==0.5.1
torchtext==0.10.0
torchvision==0.9.0+cpu
tornado==6.1
tqdm==4.64.0
traitlets==5.0.5
transformers==4.18.0
trio==0.21.0
trio-websocket==0.9.2
typer==0.3.2
typing-extensions==3.10.0.0
UniCurses==1.2
Unidecode==1.3.4
urllib3==1.26.6
uvicorn==0.18.3
virtualenv==20.7.2
visualdl==2.2.1
volumentations==0.1.8
Wand==0.6.7
wandb==0.12.11
wasabi==0.8.2
watchdog==2.1.9
wcwidth==0.2.5
webdriver-manager==3.7.0
webencodings==0.5.1
websockets==8.1
Werkzeug==2.0.3
widgetsnbextension==3.5.1
wrapt==1.15.0
wsproto==1.1.0
yacs==0.1.6
yapf==0.31.0
yarl==1.6.3
yaspin==2.1.0
zipp==3.4.1
zstandard==0.17.0
